fix : Repository 폴더 경로 정리#49
- infrastructure 가 repository 인터페이스를 구현하도록 설정
