+++ company-service/src/main/java/com/oingmaryho/business/companyservice/infrastructure/CompanyJpaRepository.java
@@ -6,8 +6,7 @@ import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.oingmaryho.business.companyservice.domain.Company;
+import com.oingmaryho.business.companyservice.domain.repository.CompanyRepository;
 
-public interface CompanyJpaRepository extends JpaRepository<Company, UUID> {
-	Optional<Company> findByIdAndIsDeletedFalse(UUID uuid);
-
+public interface CompanyJpaRepository extends JpaRepository<Company, UUID>, CompanyRepository {
 }
